feat: Actualización de formularios, lógica de aforo y panel de cliente
- Actualizados y mejorados los formularios de la aplicación.

- Añadida la lógica para calcular y asignar el aforo máximo en las sesiones.

- Implementada la nueva interfaz y lógica del panel de control para el cliente.

+++ backend/src/main/java/spotfit/modelo/dto/SesionDto.java
@@ -22,7 +22,7 @@ public class SesionDto {
     private LocalDateTime fechaInicio;
     private LocalDateTime fechaFin;
     private int aforoMaximo;
-    private int reservasActuales;
+    private long reservasActuales;
 
     public static SesionDto convertirADto(Sesion sesion) {
         SesionDto dto = new SesionDto();
